Created new analyzer script
